Add files via upload
The figure show the dataset example for different experiments considering the samples size 6, 8 and 10cm of radius. The time vs temperature for each thermocouple is presented. The blue line represents the maximum temperature corresponding to the inner sensor and the red line is the lower which correspond to external sensor.  The green line corresponds the intermedial sensor.