Add files via upload
# Table 4_ Paper in review


This repository contains experimental data and analysis related to the estimation of thermal conductivity using andesite samples from Volcan de Colima and the Eucken model (1911). The dataset includes results from various experiments performed to calculate coefficients that characterize thermal conductivity in different size of samples (6, 8 and 10 cm). The key coefficients (C and D) are accompanied by statistical measures such as the Mean Absolute Error (MAE), Root Relative Square Error (RRSE), and the coefficient of determination (R²).

## Dataset Description
The dataset consists of experimental results for multiple samples, each identified by a unique label. The following parameters are included for each sample:

- **LABEL**: Unique identifier for each sample.
- **MAE**: Mean Absolute Error, measuring the average magnitude of errors in the predictions.
- **RRSE**: Root Relative Square Error, indicating the goodness of fit of the model.
- **R²**: Coefficient of determination, representing the proportion of variance explained by the model.
- **C**: Coefficient calculated for each sample, part of the Eucken model.
- **D**: Additional coefficient calculated for each sample, also part of the Eucken model.
- **H_mean**: Mean value of the heat transfer coefficient.
- **H_error**: Error associated with the heat transfer coefficient.

## Purpose
The primary purpose of this repository is to provide a comprehensive overview of the experimental results obtained from various thermal conductivity tests. It serves as a valuable resource for researchers and practitioners in fields such as material science, thermodynamics, and engineering, enabling them to understand the thermal properties of different materials and improve the accuracy of thermal conductivity predictions.

## Files
- `experimental_data_eucken_model.xlsx`: Excel file containing the complete dataset, including all experimental results and calculated coefficients.

## Usage
To use the data:
1. Download the `experimental_data_eucken_model.xlsx` file from the repository.
2. Open it with any compatible software (e.g., Microsoft Excel, Google Sheets) to explore the dataset.
3. Utilize the data for analysis, research, or application in relevant projects.